Moved modes into their own package
