+++ src/main/java/com/github/sarhatabaot/chunkspawnerlimiter/removal/modes/EnforceKill.java
@@ -1,6 +1,7 @@
-package com.github.sarhatabaot.chunkspawnerlimiter.removal;
+package com.github.sarhatabaot.chunkspawnerlimiter.removal.modes;
 
 import com.github.sarhatabaot.chunkspawnerlimiter.chunk.ChunkCoord;
+import com.github.sarhatabaot.chunkspawnerlimiter.removal.RemovalTaskManager;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.LivingEntity;
@@ -9,35 +10,35 @@ import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.function.Consumer;
-
-public final class Kill implements RemovalMode {
+public final class EnforceKill implements RemovalMode {
     private final RemovalTaskManager removalTaskManager;
-    public Kill(RemovalTaskManager removalTaskManager) {
+    public EnforceKill(RemovalTaskManager removalTaskManager) {
         this.removalTaskManager = removalTaskManager;
     }
 
+
     @Contract(pure = true)
-    public @NotNull String getKey() { return "kill"; }
+    public @NotNull String getKey() { return "enforce-kill"; }
 
     @Override
     public void handleEntity(@NotNull Entity entity, @Nullable Cancellable event) {
-        final Consumer<Entity> action = e -> {
+        if (event != null) {
+            event.setCancelled(true);
+        }
+
+        ChunkCoord coord = ChunkCoord.from(entity.getLocation().getChunk());
+        removalTaskManager.queueChunkCheck(coord, e -> {
             if (entity instanceof LivingEntity living) {
                 living.setHealth(0);
             } else {
                 entity.remove();
             }
-        };
-
-        action.accept(entity);
-
-        ChunkCoord coord = ChunkCoord.from(entity.getLocation().getChunk());
-        removalTaskManager.queueChunkCheck(coord, action);
+        });
     }
 
     @Override
     public void handleBlock(@NotNull Block block, @NotNull Cancellable event) {
-        block.breakNaturally();
+        event.setCancelled(true);
+        //todo not sure we want to remove excess blocks, though that is what we stated we will do.
     }
 }
