+++ src/main/java/com/github/sarhatabaot/chunkspawnerlimiter/removal/modes/Remove.java
@@ -1,39 +1,26 @@
-package com.github.sarhatabaot.chunkspawnerlimiter.removal;
+package com.github.sarhatabaot.chunkspawnerlimiter.removal.modes;
 
 import com.github.sarhatabaot.chunkspawnerlimiter.chunk.ChunkCoord;
+import com.github.sarhatabaot.chunkspawnerlimiter.removal.RemovalTaskManager;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Entity;
-import org.bukkit.entity.Vehicle;
 import org.bukkit.event.Cancellable;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.function.Consumer;
-
-public final class Enforce implements RemovalMode {
+public final class Remove implements RemovalMode {
     private final RemovalTaskManager removalTaskManager;
-
-    public Enforce(RemovalTaskManager removalTaskManager) {
+    public Remove(RemovalTaskManager removalTaskManager) {
         this.removalTaskManager = removalTaskManager;
     }
 
-
     @Contract(pure = true)
-    public @NotNull String getKey() {
-        return "enforce";
-    }
-
+    public @NotNull String getKey() { return "remove"; }
 
     @Override
     public void handleEntity(@NotNull Entity entity, @Nullable Cancellable event) {
-        if (event != null) {
-            event.setCancelled(true);
-        }
-
-        if (entity instanceof Vehicle) {
-            entity.remove();
-        }
+        entity.remove();
 
         ChunkCoord coord = ChunkCoord.from(entity.getLocation().getChunk());
         removalTaskManager.queueChunkCheck(coord, Entity::remove);
@@ -41,6 +28,6 @@ public final class Enforce implements RemovalMode {
 
     @Override
     public void handleBlock(@NotNull Block block, @NotNull Cancellable event) {
-        event.setCancelled(true);
+        block.setType(org.bukkit.Material.AIR);
     }
 }
